move structure around to make more sense
