+++ .github/workflows/k8s-e2e.yaml
@@ -80,9 +80,6 @@ jobs:
           QA_AWS_REGION: ${{ secrets.QA_AWS_REGION }}
           QA_AWS_ROLE_TO_ASSUME: ${{ secrets.QA_AWS_ROLE_TO_ASSUME }}
           QA_KUBECONFIG: ${{ secrets.QA_KUBECONFIG }}
-      - name: gotestfmt
-        if: failure()
-        run: cat /tmp/gotest.log | gotestfmt
       - name: Upload test log
         uses: actions/upload-artifact@v3
         if: failure()
@@ -102,6 +99,8 @@ jobs:
       TEST_TRIGGERED_BY: chainlink-testing-framework-remote-runner-ci
     steps:
       - uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4.1.1
+      - name: Install gotestloghelper
+        run: make gotestloghelper_build
       - name: Run Remote Runner Tests
         uses: smartcontractkit/chainlink-github-actions/chainlink-testing-framework/run-tests@ce87f8986ca18336cc5015df75916c2ec0a7c4b3 # v2.1.2
         with:
@@ -116,9 +115,6 @@ jobs:
           QA_AWS_REGION: ${{ secrets.QA_AWS_REGION }}
           QA_AWS_ROLE_TO_ASSUME: ${{ secrets.QA_AWS_ROLE_TO_ASSUME }}
           QA_KUBECONFIG: ${{ secrets.QA_KUBECONFIG }}
-      - name: gotestfmt
-        if: failure()
-        run: cat /tmp/remoterunnergotest.log | gotestfmt
       - name: Upload test log
         uses: actions/upload-artifact@v2
         if: failure()
