+++ utils/clihelper/clireader.go
@@ -1,4 +1,4 @@
-package clireader
+package clihelper
 
 import (
 	"bufio"
